requirements.txt
tensorflow>=2.7.0
albumentations==1.1.0
pillow==9.0.1
tqdm==4.63.1